requirements.txt
Flask>=3.0,<4.0
requests>=2.32,<3.0
python-dotenv>=1.0,<2.0
